Pass PageEventsPopupTests
closes #522
